Minor changes to heart icon (#1)
Start with a non-filled heart (just with a colored margin) and after favoriting a dog fill the heart and reduce it's opacity.
When I was checking the app I was a bit confused with what was happening around the favorite functionality, in part due to the heart being red from the start. Maybe i'm biased because of using Instagram and that's why I was confused.
Either way, feel free to discard this if it doesn't make sense.